moved colloid links array accessors test
