--- tests/unit/test_links_arrays_accessors.c
@@ -1,135 +1,0 @@
-#include <assert.h>
-
-#include "colloids.h"
-#include "colloid_link.h"
-#include "tests.h"
-
-int test_links_arrays_accessors(void);
-int test_links_arrays_accessors_with_state(void);
-
-/*****************************************************************************
- *
- *  test_links_arrays_accessors_suite
- *
- *****************************************************************************/
-
-int test_links_arrays_accessors_suite(void) {
-  test_links_arrays_accessors();
-  test_links_arrays_accessors_with_state();
-  
-  return 0;
-}
-
-/*****************************************************************************
- *
- *  test_links_arrays_accessors
- *
- *****************************************************************************/
-
-int test_links_arrays_accessors(void) {
-  pe_t * pe = NULL;
-  cs_t * cs = NULL;
-  colloids_info_t * cinfo = NULL;
-  colloid_options_t opts = colloid_options_default();
-  colloid_t * pc;
-  double r[3] = {0.5, 0.5, 0.5};
-
-  pe_create(MPI_COMM_WORLD, PE_QUIET, &pe);
-  cs_create(pe, &cs);
-  cs_init(cs);
-  colloids_info_create(pe, cs, &opts, &cinfo);
-  colloids_info_add_local(cinfo, 1, r, 1.0, &pc);
-  pc->lnk = colloid_link_allocate();
-  pc->lnk->next = NULL;
-  
-  pc->lnk->i = 3;
-  pc->lnk->j = 2;
-  pc->lnk->p = 1;
-  pc->lnk->status = 0;
-  for (int i = 0; i < 3; i++)
-    pc->lnk->rb[i] = i;
-
-  assert(pc->links != NULL);
-  assert(pc);
-  assert(pc->lnk);
-  assert(pc->lnk->next == NULL);
-  copy_links_to_array(pc);
-
-  int i, j, p, status;
-  double rb[3];
-
-  colloid_link_i(pc->links, 0, &i);
-  colloid_link_j(pc->links, 0, &j);
-  colloid_link_p(pc->links, 0, &p);
-  colloid_link_status(pc->links, 0, &status);
-  colloid_link_rb(pc->links, 0, rb);
-
-  test_assert(i == pc->lnk->i);
-  test_assert(j == pc->lnk->j);
-  test_assert(p == pc->lnk->p);
-  test_assert(status == pc->lnk->status);
-  for (int index = 0; index < 3; index++)
-    test_assert(rb[index] == pc->lnk->rb[index]);
-
-  pe_info(pe, "PASS     ./unit/test_links_arrays_accessors\n");
-
-  return 0;
-}
-
-/*****************************************************************************
- *
- *  test_links_arrays_accessors_with_state
- *
- *****************************************************************************/
-
-int test_links_arrays_accessors_with_state(void) {
-  pe_t * pe = NULL;
-  cs_t * cs = NULL;
-  colloids_info_t * cinfo = NULL;
-  colloid_options_t opts = colloid_options_default();
-  colloid_t * pc;
-  colloid_state_t state;
-  double r[3] = {0.5, 0.5, 0.5};
-
-  pe_create(MPI_COMM_WORLD, PE_QUIET, &pe);
-  cs_create(pe, &cs);
-  cs_init(cs);
-  colloids_info_create(pe, cs, &opts, &cinfo);
-  create_dummy_state(&state, 1, 1.0, r);
-  colloids_info_add_local_with_state(cinfo, &state, &pc);
-  pc->lnk = colloid_link_allocate();
-  pc->lnk->next = NULL;
-  
-  pc->lnk->i = 3;
-  pc->lnk->j = 2;
-  pc->lnk->p = 1;
-  pc->lnk->status = 0;
-  for (int i = 0; i < 3; i++)
-    pc->lnk->rb[i] = i;
-
-  assert(pc->links != NULL);
-  assert(pc);
-  assert(pc->lnk);
-  assert(pc->lnk->next == NULL);
-  copy_links_to_array(pc);
-
-  int i, j, p, status;
-  double rb[3];
-
-  colloid_link_i(pc->links, 0, &i);
-  colloid_link_j(pc->links, 0, &j);
-  colloid_link_p(pc->links, 0, &p);
-  colloid_link_status(pc->links, 0, &status);
-  colloid_link_rb(pc->links, 0, rb);
-
-  test_assert(i == pc->lnk->i);
-  test_assert(j == pc->lnk->j);
-  test_assert(p == pc->lnk->p);
-  test_assert(status == pc->lnk->status);
-  for (int index = 0; index < 3; index++)
-    test_assert(rb[index] == pc->lnk->rb[index]);
-
-  pe_info(pe, "PASS     ./unit/test_links_arrays_accessors\n");
-
-  return 0;
-}
